+++ app/Http/Controllers/JenjangPendidikanController.php
@@ -9,53 +9,72 @@ class JenjangPendidikanController extends Controller
 {
     public function index()
     {
-        $pendidikan = JenjangPendidikan::all();
-        return view('pelengkap.pendidikan.index', compact('pendidikan'));
+        if (user()?->hasRole('admin')) {
+            $pendidikan = JenjangPendidikan::all();
+            return view('pelengkap.pendidikan.index', compact('pendidikan'));
+        } else {
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     public function create()
     {
-        return view('pelengkap.pendidikan.create');
+        if (user()?->hasRole('admin')) {
+            return view('pelengkap.pendidikan.create');
+        } else {
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     public function store(Request $request)
     {
-        $validated = $request->validate([
-            'nama_jenjang_pendidikan' => 'required',
-        ], [
-            'nama_jenjang_pendidikan.required' => 'Nama pendidikan wajib diisi.',
-        ]);
-        JenjangPendidikan::create($validated);
-        return redirect()->route('jenjang-pendidikan.index')->with('success', 'pendidikan berhasil disimpan');        
-    }
-
-    public function show($id)
-    {
-        //
+        if (user()?->hasRole('admin')) {
+            $validated = $request->validate([
+                'nama_jenjang_pendidikan' => 'required',
+            ], [
+                'nama_jenjang_pendidikan.required' => 'Nama pendidikan wajib diisi.',
+            ]);
+            JenjangPendidikan::create($validated);
+            return redirect()->route('jenjang-pendidikan.index')->with('success', 'pendidikan berhasil disimpan'); 
+        } else {
+            return response()->view('errors.403', [abort(403)], 403);
+        }       
     }
 
     public function edit($id)
     {
-        $pendidikan = JenjangPendidikan::findOrFail($id);
-        return view('pelengkap.pendidikan.edit', compact('pendidikan'));
+        if (user()?->hasRole('admin')) {
+            $pendidikan = JenjangPendidikan::findOrFail($id);
+            return view('pelengkap.pendidikan.edit', compact('pendidikan'));
+        } else {
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     public function update(Request $request, $id)
     {
-        $validated = $request->validate([
-            'nama_jenjang_pendidikan' => 'required',
-        ], [
-            'nama_jenjang_pendidikan.required' => 'Nama pendidikan wajib diisi.',
-        ]);
-        $pendidikan = JenjangPendidikan::findOrFail($id);
-        $pendidikan->update($validated);
-        return redirect()->route('jenjang-pendidikan.index')->with('success', 'pendidikan berhasil diupdate');
+        if (user()?->hasRole('admin')) {
+            $validated = $request->validate([
+                'nama_jenjang_pendidikan' => 'required',
+            ], [
+                'nama_jenjang_pendidikan.required' => 'Nama pendidikan wajib diisi.',
+            ]);
+            $pendidikan = JenjangPendidikan::findOrFail($id);
+            $pendidikan->update($validated);
+            return redirect()->route('jenjang-pendidikan.index')->with('success', 'pendidikan berhasil diupdate');
+        } else {
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     public function destroy($id)
     {
-        $pendidikan = JenjangPendidikan::findOrFail($id);
-        $pendidikan->delete();
-        return redirect()->route('jenjang-pendidikan.index')->with('success', 'pendidikan berhasil dihapus');
+        if (user()?->hasRole('admin')) {
+            $pendidikan = JenjangPendidikan::findOrFail($id);
+            $pendidikan->delete();
+            return redirect()->route('jenjang-pendidikan.index')->with('success', 'pendidikan berhasil dihapus');
+        } else {
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 }
